requirements.txt
# Core
arrow==0.10.0
fastjsonschema==1.6
gunicorn==19.4.5
jinja2==2.8.1
jsonschema==2.6.0
neo4j-driver==1.6.1
neo4jrestclient==2.1.1
numpy==1.12.1
pandas==0.23.3
pycrypto==2.6.1
pyyaml==3.11
redis==2.10.5
requests==2.9.1
werkzeug==0.14.1

# Apache Airflow
apache-airflow==1.9.0
cryptography==2.4.2

# SqlAlchemy
alembic==0.8.10
psycopg2-binary==2.7.5
sqlalchemy==1.1.0
sqlalchemy-utils==0.33.10

# Flask
flask==0.11.1
flask-admin==1.4.1
flask-cors==2.1.2
flask-jsonschema==0.1.1
flask-login==0.2.11
flask-migrate==2.2.1
flask-script==2.0.6
flask-sqlalchemy==2.1

# Celery
celery==4.1.1
flower==0.9.1

# Kafka
kafka-python==1.4.3

# Documentation
sphinx==1.8.3
sphinx-rtd-theme==0.4.2
sphinxcontrib-mermaid==0.3.1
sphinxcontrib-httpdomain==1.7.0

# Tests
pytest==4.1.0
deepdiff==3.3.0
pytest-cov==2.6.1
pytest-freezegun==0.3.0.post1
pytest-mock==1.10.0
